==> step_01_merge_GAEZ_to_NC.py <==
import numpy as np
import pandas as pd
import xarray as xr
import rioxarray as rxr


# Read GAEZ dataframe, which contains mapping of path to tif files
#   by default, input_level = High
GAEZ_df = pd.read_csv('data/GAEZ_v4/GAEZ_df.csv').query('input_level == "High"')
GAEZ_4_vars = ["year", "model", "rcp", "crop", "water_supply", "c02_fertilization"]


year_rename = {
    '1981-2010':1995,
    '2011-2040':2025,
    '2041-2070':2055,
    '2071-2100':2085
}



# helpfer function to convert dataframe to xarray DataArray
def get_xr_darray_from_df(in_df):
    arrs = [] 
    for idx,row in in_df.iterrows():
        dim_coord = {k:[v] for k,v in zip(in_df.index.names, idx)}
        arrs.append(
            rxr.open_rasterio(row['fpath'] + '_clipped.tif', masked=True)
            .drop_vars('band')
            .squeeze()
            .expand_dims(dim_coord)
        )

    out_xr = xr.combine_by_coords(arrs, combine_attrs='drop')
    return out_xr

def calc_mean_std(in_xr):
    mean_xr = in_xr.mean(dim='model').expand_dims(band=['mean'])
    std_xr = in_xr.std(dim='model').expand_dims(band=['std'])
    return xr.concat([mean_xr, std_xr], dim='band')

# Get historical yield
#  set the standard deviation for historical yield is 0
#  this make it compatible with future yield data structure
GAEZ_4_hist = (
    GAEZ_df.query('gaez_cat == "GAEZ_4" and name.str.contains("ycHa|ycHg") and rcp == "Historical"')
    .reset_index(drop=True)[[*GAEZ_4_vars, 'fpath']]
    .replace({'year': year_rename})
    .set_index(GAEZ_4_vars)
    .reset_index(['model', 'rcp', 'c02_fertilization'], drop=True)
)

GAEZ_4_hist_xr = get_xr_darray_from_df(GAEZ_4_hist)
GAEZ_4_hist_xr = xr.concat(
    [GAEZ_4_hist_xr.expand_dims(band=['mean']),
     (GAEZ_4_hist_xr * 0).expand_dims(band=['std'])],
    dim='band'
)


# Get future yield
#  Rice has missing rows:
#   year=2071-2100, model=MIROC-ESM-CHEM, rcp=RCP4.5, crop=Wetland rice, water_supply=Irrigated, co2_fertilization=With CO2 Fertilization
#   year=2071-2100, model=MIROC-ESM-CHEM, rcp=RCP4.5, crop=Wetland rice, water_supply=Irrigated, co2_fertilization=Without CO2 Fertilization

GAEZ_4_future = (
    GAEZ_df.query('gaez_cat == "GAEZ_4" and rcp != "Historical"')
    .replace({'year': year_rename})
    .reset_index(drop=True)[[*GAEZ_4_vars, 'fpath']]
)
GAEZ_4_future_wheat = GAEZ_4_future.query('crop == "Wheat"').set_index(GAEZ_4_vars)
GAEZ_4_future_maize = GAEZ_4_future.query('crop == "Maize"').set_index(GAEZ_4_vars)
GAEZ_4_future_rice = GAEZ_4_future.query('crop == "Wetland rice"').set_index(GAEZ_4_vars)

GAEZ_4_future_wheat_xr = calc_mean_std(get_xr_darray_from_df(GAEZ_4_future_wheat))
GAEZ_4_future_maize_xr = calc_mean_std(get_xr_darray_from_df(GAEZ_4_future_maize))
GAEZ_4_future_rice_xr = calc_mean_std(get_xr_darray_from_df(GAEZ_4_future_rice))

GAEZ_4_future_xr = xr.concat(
    [GAEZ_4_future_wheat_xr, GAEZ_4_future_maize_xr, GAEZ_4_future_rice_xr],
    dim='crop'
)


# Convert (kg dry-weight)/ha to (t yield)/ha
#   Conversion factor from Michalis
convesion_factor = xr.DataArray(
    np.array([0.87 * 1e-3, 0.875 * 1e-3, 0.875 * 1e-3], dtype=np.float32), 
    dims=['crop'], 
    coords={'crop': ['Maize', 'Wetland rice', 'Wheat']}
)

GAEZ_4_hist_xr = GAEZ_4_hist_xr * convesion_factor
GAEZ_4_future_xr = GAEZ_4_future_xr * convesion_factor

# Save to netcdf
GAEZ_4_hist_xr.to_netcdf('data/GAEZ_v4/GAEZ_4_historical_t_ha.nc')
GAEZ_4_future_xr.to_netcdf('data/GAEZ_v4/GAEZ_4_future_t_ha.nc')
